Mobassher

+++ src/app/modules/projects/project.service.ts
@@ -1,3 +1,6 @@
+import httpStatus from 'http-status'
+import QueryBuilder from '../../builder/QueryBuilder'
+import AppError from '../../errors/AppError'
 import { IProject } from './project.interface'
 import { Project } from './project.model'
 
@@ -11,10 +14,20 @@ const createProject = async (payload: IProject) => {
   return result
 }
 
-const getAllProjects = async () => {
-  const result = await Project.find().sort('position')
-  // console.log(result)
-  return result
+const getAllProjects = async (query: Record<string, unknown>) => {
+  const productQuery = new QueryBuilder(Project.find(), query)
+    .search(['title'])
+    .filter()
+    .sort()
+    .paginate()
+    .fields()
+
+  const result = await productQuery.modelQuery
+  const meta = await productQuery.countTotal()
+  return {
+    meta,
+    result,
+  }
 }
 
 const getSingleProject = async (id: string) => {
@@ -23,7 +36,10 @@ const getSingleProject = async (id: string) => {
 }
 
 const updateProject = async (id: string, payload: Partial<IProject>) => {
-  // console.log(payload)
+  const isProductExist = await Project.findById(id)
+  if (!isProductExist) {
+    throw new AppError(httpStatus.NOT_FOUND, 'This Product is not found')
+  }
   const result = await Project.findByIdAndUpdate(id, payload, {
     new: true,
     runValidators: true,
@@ -33,6 +49,10 @@ const updateProject = async (id: string, payload: Partial<IProject>) => {
 }
 
 const deleteProject = async (id: string) => {
+  const isProductExist = await Project.findById(id)
+  if (!isProductExist) {
+    throw new AppError(httpStatus.NOT_FOUND, 'This Product is not found')
+  }
   await Project.findByIdAndDelete(id)
   return null
 }
